HDX-9781 Start moving view as table definitions and tests to two separate files
