+++ src/hapi_schema/db_admin1.py
@@ -64,20 +64,3 @@ view_params_admin1 = ViewParams(
         )
     ),
 )
-
-
-class DBAdmin1VAT(Base):
-    __tablename__ = "admin1_vat"
-    id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    location_ref: Mapped[int] = mapped_column(Integer)
-    code: Mapped[str] = mapped_column(String(128), index=True)
-    name: Mapped[str] = mapped_column(String(512), index=True)
-    is_unspecified: Mapped[bool] = mapped_column(Boolean)
-    reference_period_start: Mapped[datetime] = mapped_column(
-        DateTime, index=True
-    )
-    reference_period_end: Mapped[datetime] = mapped_column(
-        DateTime, index=True
-    )
-    location_code: Mapped[str] = mapped_column(String(128), index=True)
-    location_name: Mapped[str] = mapped_column(String(512), index=True)
