+++ src/hapi_schema/db_views_as_tables.py
@@ -1,0 +1,28 @@
+from datetime import datetime
+
+from sqlalchemy import (
+    Boolean,
+    DateTime,
+    Integer,
+    String,
+)
+from sqlalchemy.orm import Mapped, mapped_column
+
+from hapi_schema.utils.base import Base
+
+
+class DBAdmin1VAT(Base):
+    __tablename__ = "admin1_vat"
+    id: Mapped[int] = mapped_column(Integer, primary_key=True)
+    location_ref: Mapped[int] = mapped_column(Integer)
+    code: Mapped[str] = mapped_column(String(128), index=True)
+    name: Mapped[str] = mapped_column(String(512), index=True)
+    is_unspecified: Mapped[bool] = mapped_column(Boolean)
+    reference_period_start: Mapped[datetime] = mapped_column(
+        DateTime, index=True
+    )
+    reference_period_end: Mapped[datetime] = mapped_column(
+        DateTime, index=True
+    )
+    location_code: Mapped[str] = mapped_column(String(128), index=True)
+    location_name: Mapped[str] = mapped_column(String(512), index=True)
